use Dial from pkg/psmdb/mongo
